only delete files if folder exists

+++ src/extract-dose-history.py
@@ -53,7 +53,7 @@ def updateZipCodeFilesForDrug(localBasePath, drugs):
 
   for drug in drugs:
       drugPath = localBasePath + 'data/therapeutics/' + drug.lower() + '/'
-      if lastProcessedDate < "2020-03-02T00:00:00.000":
+      if lastProcessedDate < "2020-03-02T00:00:00.000" and os.path.exists(drugPath):
         for root, directories, files in os.walk(drugPath):
           for file in files:
             os.remove(os.path.join(root, file))
